renamed the gopigo3frc directory to frc, added the path gen tool

+++ gopigo3/gopigo3navx/InertialDataIntegrator.cpp
@@ -5,7 +5,7 @@
  *      Author: Scott
  */
 
-#include <InertialDataIntegrator.h>
+#include "InertialDataIntegrator.h"
 
 InertialDataIntegrator::InertialDataIntegrator() {
     ResetDisplacement();
@@ -17,7 +17,7 @@ void InertialDataIntegrator::UpdateDisplacement( float accel_x_g, float accel_y_
         float accel_g[2];
         float accel_m_s2[2];
         float curr_velocity_m_s[2];
-        float sample_time = (1.0f / static_cast<float>(update_rate_hz));
+        float sample_time = (1.0f / update_rate_hz);
         accel_g[0] = accel_x_g;
         accel_g[1] = accel_y_g;
         for ( int i = 0; i < 2; i++ ) {
